requirements.txt
Flask==2.0.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
email-validator==1.1.3
Werkzeug==2.0.1
Jinja2==3.0.1
SQLAlchemy==1.4.23
Flask-SQLAlchemy==2.5.1
bcrypt==3.2.0
# Advanced features
Flask-Mail==0.9.1
Flask-SocketIO==5.1.1
scikit-learn==1.0.2
nltk==3.6.5
pandas==1.3.4
matplotlib==3.5.0
gunicorn==20.1.0
celery==5.1.2
python-dotenv==0.19.2
requests==2.26.0
# Resume parsing
google-generativeai==0.3.1 